Update Employee Card styling (#159)
* Init

* Fix attribute splitting

* Snapshots

* Fix center arrow

* Tweak

* switch total results to use total count (#161)

* switch total results to use total count

* fix snapshot test

* address pr comments
