+++ components/types.ts
@@ -109,6 +109,7 @@ export enum MeetingType {
 // Represents the course and employee data returned by /search
 export interface SearchResult {
   results: SearchItem[];
+  totalCount: number;
   filterOptions: FilterOptions;
   hasNextPage: boolean;
 }
@@ -124,6 +125,7 @@ export type SearchItem = CourseResult | Employee;
 export function BLANK_SEARCH_RESULT(): SearchResult {
   return {
     results: [],
+    totalCount: 0,
     filterOptions: {
       nupath: [],
       subject: [],
