+++ styles/base.scss
@@ -10,7 +10,6 @@
 @import 'panels/BaseClassPanel';
 @import 'panels/DesktopClassPanel';
 @import 'panels/DesktopSectionPanel';
-@import 'panels/EmployeePanel';
 @import 'panels/MobileClassPanel';
 @import 'panels/MobileSectionPanel';
 @import 'panels/SplashPage';
