+++ styles/results/_MobileSearchResult.scss
@@ -10,6 +10,7 @@
     border-radius: 8px;
     padding: 16px;
     display: flex;
+    align-items: center;
     cursor: pointer;
 
     &:focus,
@@ -25,6 +26,7 @@
       border-bottom: none;
       padding: 16px;
       display: flex;
+      align-items: center;
 
       &:focus,
       &:active {
@@ -52,6 +54,23 @@
       color: #505050;
       margin-left: 24px;
     }
+
+    &--employeeName {
+      font-style: normal;
+      font-weight: 800;
+      font-size: 20px;
+      line-height: 24px;
+      display: flex;
+      align-items: center;
+      color: #505050;
+      margin-left: 24px;
+    }
+  }
+
+  &__employeeBody {
+    font-size: 16px;
+    line-height: 20px;
+    padding: 10px 30px;
   }
 
   &__panel {
