+++ styles/results/_SearchResult.scss
@@ -29,6 +29,12 @@
       border-radius: 5px 5px 0 0;
     }
 
+    &--employeeName {
+      margin-bottom: 4px;
+      font-size: 18px;
+      border-radius: 5px 5px 0 0;
+    }
+
     &--creditString {
       font-size: 16px;
       color: #474747;
@@ -67,6 +73,13 @@
   &__panel {
     padding: 16px 16px 32px 16px;
     font-size: 16px;
+
+    &--column {
+      display: inline-block;
+      min-width: 300px;
+      padding-right: 50px;
+      vertical-align: top;
+    }
 
     &--main {
       display: flex;
